Doubles
Doubles on GPS

+++ src/com/F21AS_CW/GPSCoordinates.java
@@ -1,19 +1,19 @@
 package com.F21AS_CW;
 
 public class GPSCoordinates {
-    private int longitude;
-    private int latitude;
+    private double longitude;
+    private double latitude;
 
-    public GPSCoordinates(int l0ng, int lat) {
+    public GPSCoordinates(double l0ng, double lat) {
         longitude = l0ng;
         latitude = lat;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
